Add email field validation

+++ src/handlers/registerNotification.js
@@ -4,6 +4,7 @@ const serverless = require('serverless-http');
 const bodyParser = require('body-parser');
 const app = express()
 const sendEmail = require('../sendEmail')
+const emailValidator = require('email-validator')
 
 
 app.use(bodyParser.urlencoded({ extended: true })); 
@@ -11,31 +12,39 @@ app.use(bodyParser.urlencoded({ extended: true }));
 app.post('/register', function (req, res) {
   getDatabase().then(db => {
     db.containers.createIfNotExists({ id: "notifications" }).then(containerResponse => {
-      // TODO actuial verification and parsing out the body
-      // TODO delete if no stores selected
-      let emailAddress = req.body.email
+      // SES and Cosmos both care about the casing, so lets always toLower it
+      let emailAddress = req.body.email.toLowerCase()
+      if(emailValidator.validate(emailAddress)){
+        let notification = {}
+        notification.walmartStores = req.body.walmartStores
+        notification.walgreensStores = req.body.walgreensStores
+        notification.krogerStores = req.body.krogerStores
+        notification.albertsonsStores = req.body.albertsonsStores
 
-      let notification = req.body
-      notification.id = emailAddress
+        notification.id = emailAddress
 
-      containerResponse.container.items.upsert(notification).then(itemResponse => {
-        let sendPromise = null
-        if (itemResponse.statusCode == 201){
-          sendPromise = sendEmail(emailAddress,"Verify email address for covid vaccine notifications", `Click this link to confirm you would like to recieve covid vaccine notifications for the following stores.`)
-        } else {
-          sendPromise = sendEmail(emailAddress,"Your covid vaccine notification preferences have been updated.","More details")
-        }
-        // Handle promise's fulfilled/rejected states
-        sendPromise.then(
-          function(data) {
-            console.log(data.MessageId);
-            res.send("Success");
-          }).catch(
-            function(err) {
-              console.error(err, err.stack);
-              res.send("Failed");
-          });
-      });
+        containerResponse.container.items.upsert(notification).then(itemResponse => {
+          let sendPromise = null
+          if (itemResponse.statusCode == 201){
+            sendPromise = sendEmail(emailAddress,"You have signed up for covid vaccine appointment notifications.", `You will be notified when any appointments become available at the pharmacies you selected.`
+            )
+          } else {
+            sendPromise = sendEmail(emailAddress,"Your covid vaccine notification preferences have been updated.","Your pharmacy selections have been updated for covid vaccine appointment notifications.")
+          }
+          // Handle promise's fulfilled/rejected states
+          sendPromise.then(
+            function(data) {
+              console.log(data.MessageId);
+              res.send("Success");
+            }).catch(
+              function(err) {
+                console.error(err, err.stack);
+                res.send("Failed");
+            });
+       });
+      } else {
+        res.status(400).send("Invalid request.")
+      }
     });
   });
 })
@@ -43,31 +52,33 @@ app.post('/register', function (req, res) {
 app.get('/unregister', function (req, res) {
   getDatabase().then(db => {
     db.containers.createIfNotExists({ id: "notifications" }).then(containerResponse => {
-      // TODO actuial verification and parsing out the body
-      // TODO delete if no stores selected
-      let emailAddress = req.query.email
-
-      containerResponse.container.item(emailAddress).delete().then(() => {
-        console.log(`Unsubscribing ${emailAddress}`)
-        let sendPromise = sendEmail(emailAddress,"Unsubscribed from covid vaccine notifications", `You have been unsubscribed from covid vaccine notifications.`)
-        
-        // Handle promise's fulfilled/rejected states
-        sendPromise.then(
-          function(data) {
-            console.log(data.MessageId);
-            res.send("Success");
-          }).catch(
-            function(err) {
+      // SES and Cosmos both care about the casing, so lets always toLower it
+      let emailAddress = req.query.email.toLowerCase()
+      if(emailValidator.validate(emailAddress)){
+        containerResponse.container.item(emailAddress).delete().then(() => {
+          console.log(`Unsubscribing ${emailAddress}`)
+          let sendPromise = sendEmail(emailAddress,"Unsubscribed from covid vaccine notifications", `You have been unsubscribed from covid vaccine appointmnent notifications.`)
+          
+          // Handle promise's fulfilled/rejected states
+          sendPromise.then(
+            function(data) {
+              console.log(data.MessageId);
+              res.send("Success");
+            }).catch(
+              function(err) {
+              console.error(err, err.stack);
+              res.send("Something went wrong!");
+            });
+        }).catch(
+          function(err) {
+            // Always send success, regardless of if the record was found and deleted
             console.error(err, err.stack);
             res.send("Success");
-          });
-      }).catch(
-        function(err) {
-          // Always send success, regardless of if the record was found and deleted
-          console.error(err, err.stack);
-          res.send("Success");
-        }
-      )
+          }
+        )
+      } else {
+        res.status(400).send("Invalid request.")
+      }
     });
   });
 })
